Share gallery preview card with AutoCull
